Add argument count printing functionality
Implements the main function to print the number of arguments passed into the program. The function takes in argc (argument count) and argv (array of strings containing the arguments). Since we are only interested in the argument count, we ignore the argv parameter by using (void)argv. The program then prints argc - 1 to exclude the program name itself. This functionality is essential for displaying the number of arguments passed to the program.