Implement free_grid function
Implements the free_grid function that frees a 2D grid previously created by the alloc_grid function. The function takes in grid as the grid to be freed and height as the height of the grid. It first checks if grid is NULL or if height is less than or equal to 0, in which case it returns without performing any further actions. Otherwise, it proceeds to free the memory allocated for each row of the grid using a loop. Finally, it frees the memory allocated for the grid itself. This ensures that all the memory associated with the grid is properly deallocated.