Implement string_nconcat function
Implements the string_nconcat function that concatenates two strings. The function takes in s1 as the first string, s2 as the second string, and n as the number of bytes from s2 to concatenate. If s1 or s2 is NULL, it is treated as an empty string. The function calculates the lengths of s1 and s2 using strlen and determines the length of the resulting concatenated string cat. If n is greater than or equal to the length of s2, it is adjusted to the length of s2. Then, the function allocates memory for cat using malloc and copies the characters from s1 and the first n characters from s2 into cat. Finally, the null-terminating character is added at the end of cat. If malloc fails to allocate memory, the function returns NULL. Otherwise, it returns the pointer to the concatenated string cat. This function provides a convenient way to concatenate strings while limiting the number of bytes from the second string.