Implement array_range function
Implements the array_range function that creates an array of integers. The function takes in min as the minimum value (inclusive) and max as the maximum value (inclusive). If min is greater than max, the function returns NULL. Otherwise, it calculates the size of the array by subtracting min from max and adding 1. Then, it calls malloc to allocate memory for the array and checks if malloc fails. If malloc succeeds, the function uses a loop to populate the array with values from min to max. Finally, it returns the pointer to the newly created array. This function provides a convenient way to create and initialize an array of integers with a specified range.