Add function to add a node at the end of a listint_t list
Implement the add_nodeint_end function, which adds a new node with the given value at the end of a listint_t list. The function allocates memory for the new node, assigns the value, and sets the next pointer to NULL. If the list is empty, it updates the head to point to the new node. Otherwise, it traverses the list to find the last node and updates its next pointer to the new node. Finally, it returns the address of the new node. If memory allocation fails, it returns NULL.