Add function to sum the data of a listint_t list
Implement the sum_listint function, which returns the sum of all the data (n) in a listint_t linked list. The function takes a pointer to the head of the list as an argument. It initializes a sum variable to 0 and a current pointer to the head. It then iterates through the list, adding the value of each node's data to the sum, until it reaches the end of the list. Finally, it returns the sum. This function provides a convenient way to calculate the total sum of the data values stored in the list.