Add function to get bit value at index
Implement the get_bit function, which returns the value of a bit at a given index in an unsigned long int number. The function first checks if the index is greater than or equal to the number of bits in an unsigned long int. If so, it returns -1 to indicate an error. Otherwise, it right-shifts the number by the index and performs a bitwise AND operation with 1 to extract the bit at the specified index. The result is returned as the value of the bit. This function provides a convenient way to retrieve the value of a specific bit in a number.