Add function to set bit value at index
Implement the set_bit function, which sets the value of a bit to 1 at a given index in an unsigned long int number. The function first checks if the index is greater than or equal to the number of bits in an unsigned long int. If so, it returns -1 to indicate an error. Otherwise, it uses a bitwise OR operation to set the bit at the specified index to 1. The updated value is stored back in the original variable using a pointer. The function returns 1 to indicate success. This function provides a convenient way to modify a specific bit in a number.