Add function to clear bit value at index
Implement the clear_bit function, which sets the value of a bit to 0 at a given index in an unsigned long int number. The function first checks if the index is greater than or equal to the number of bits in an unsigned long int. If so, it returns -1 to indicate an error. Otherwise, it uses a bitwise NOT and AND operation to clear the bit at the specified index by creating a mask with 0 at the index position and 1 everywhere else. The updated value is stored back in the original variable using a pointer. The function returns 1 to indicate success. This function provides a convenient way to clear a specific bit in a number.