Add function to calculate flipped bits count
Implement the flip_bits function, which calculates the number of bits that need to be flipped to get from one number to another. The function takes two unsigned long int numbers as input. It uses the XOR operator to perform a bitwise XOR operation between the two numbers, resulting in a number that has 1s in the positions where the bits are different between the two numbers. It then iterates over the XOR result, counting the number of 1 bits using a bitwise AND operation with 1 and shifting the bits to the right. The function returns the count of flipped bits. This function provides a way to determine the Hamming distance between two numbers, which represents the number of positions at which the corresponding bits are different.