Print and count elements of a doubly linked list
Implemented 'print_dlistint' function to print all elements of a doubly linked list and count the number of nodes. The function iterates through the list, printing the value of each node and moving to the next node until the end of the list is reached. The count of nodes is updated as each node is processed.