Inserting node at a given position in a doubly linked list
Implemented 'insert_dnodeint_at_index' function in 'lists.h' for inserting a new node at a given index in a doubly linked list. The function takes a double pointer to the head of the list, the index, and the value to be assigned to the new node. It handles cases for inserting at the beginning and within the list, updating the previous and next pointers accordingly.