Implement hash_table_create function
Implemented the 'hash_table_create' function in the 'hash_tables.c' file. This function creates a new hash table of the specified size and initializes its array elements to NULL. Memory allocation for the hash table and its array is also handled, with appropriate error checking and cleanup in case of failure.