Add Hash Table Header
Added the 'hash_tables.h' header file, which defines the structures and function prototypes required for implementing a hash table. This header includes definitions for 'hash_node_t' and 'hash_table_t' structs, as well as function prototypes for hash table operations. This header will be used for implementing and working with hash tables in the project.