Implement DJB2 Hash Function
Added the 'hash_djb2' function to the hash_tables.c file. This function implements the DJB2 hash algorithm, which is used for hashing strings in hash tables. The DJB2 algorithm iterates through the input string and calculates a hash value. This function is an essential part of the hash table implementation, as it provides the hash value for keys when inserting or retrieving data in the hash table.