Implement Key Index Function
Added the 'key_index' function to the hash_tables.c file. This function calculates the index at which a key/value pair should be stored in the array of a hash table. It uses the DJB2 hash algorithm implemented in the hash_djb2 function. The 'key_index' function takes the key and the size of the hash table array as parameters and returns the calculated index. This function is a fundamental part of hash table operations, as it determines where data should be stored and retrieved within the hash table.