Added support for building with different compiler optimization levels
The previous Makefile only allowed the school executable to be built with the default compiler optimization level. This commit updates the Makefile to allow for different compiler optimization levels to be specified using the 'OPT' variable. For example, to build the executable with the highest optimization level, you would run 'make OPT=-O3'.